requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.2
django-countries==7.2.1
django-localflavor==3.1
djangorestframework==3.13.1
frozenlist==1.3.0
geoip2==4.5.0
gunicorn==20.1.0
idna==3.3
maxminddb==2.2.0
multidict==6.0.2
python-stdnum==1.17
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==6.0.0
yarl==1.7.2
